test: move u32_ops proc tests to nested module structure for readability
